requirement.txt
django==2.1.7
DjangoUeditor==1.8.143
django-haystack==2.6.1
jieba==0.38
Whoosh==2.7.4
djangorestframework==3.8.0
django-rest-swagger==2.1.1
django_filter==2.2.0
djangorestframework-jwt==1.11.0
xadmin2==2.0.1
pyquery==1.4.0
Markdow==3.1.1
django-ckeditor==5.7.1
django-mptt==0.10.7
django-markdown==0.8.4
Pillow==6.2.0
PyMySQL==0.9.3
django2-markdown==0.8.6
